+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Changed
 * Change to finish input with a newline when enter is pressed inside the input line
+* Change Reader to stop listening to key events
 
 ### Fixed
 * Fix navigation to the previous history line to display the latest line